+++ contracts/multisig-prover/src/encoding/bcs.rs
@@ -4,9 +4,7 @@ use axelar_wasm_std::hash::Hash;
 use cosmwasm_std::HexBinary;
 use error_stack::{Result, ResultExt};
 use k256::ecdsa::RecoveryId;
-use multisig::key::Signature;
-use multisig::msg::SignerWithSig;
-use multisig::verifier_set::VerifierSet;
+use multisig::{Signature,SignerWithSig,VerifierSet};
 use sha3::{Digest, Keccak256};
 use sui_gateway::{CommandType, ExecuteData, Message, MessageToSign, Proof, WeightedSigners};
