+++ contracts/multisig-prover/src/encoding/stellar_xdr.rs
@@ -2,8 +2,7 @@ use axelar_wasm_std::hash::Hash;
 use axelar_wasm_std::FnExt;
 use cosmwasm_std::HexBinary;
 use error_stack::{Result, ResultExt};
-use multisig::msg::SignerWithSig;
-use multisig::verifier_set::VerifierSet;
+use multisig::{SignerWithSig,VerifierSet};
 use sha3::{Digest, Keccak256};
 use stellar::{Message, Messages, Proof, WeightedSigners};
 use stellar_xdr::curr::{Limits, ScVal, WriteXdr};
