+++ contracts/multisig/src/exported.rs
@@ -1,0 +1,6 @@
+
+pub use crate::msg::{ Signer,SignerWithSig};  
+pub use crate::key::{KeyType, PublicKey, Signature};  
+pub use crate::verifier_set::VerifierSet;  
+pub use crate::multisig::Multisig;  
+pub use crate::types::{MultisigState,MsgToSign};
