+++ contracts/voting-verifier/src/state.rs
@@ -6,7 +6,7 @@ use axelar_wasm_std::{counter, nonempty, MajorityThreshold};
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Order, StdResult, Storage};
 use cw_storage_plus::{Index, IndexList, IndexedMap, Item, Map, MultiIndex};
-use multisig::verifier_set::VerifierSet;
+use multisig::VerifierSet;
 use router_api::{ChainName, Message};
 
 use crate::error::ContractError;
